reorg bc circular dep
